+++ backend/src/main/java/br/com/estacionamento/model/Localidade.java
@@ -3,15 +3,16 @@ package br.com.estacionamento.model;
 import java.io.Serializable;
 import javax.persistence.*;
 import java.util.List;
+
 /**
  * The persistent class for the localidade database table.
  * 
  */
 @Entity
-@NamedQuery(name="Localidade.findAll", query="SELECT l FROM Localidade l")
+@NamedQuery(name = "Localidade.findAll", query = "SELECT l FROM Localidade l")
 public class Localidade implements Serializable {
 	private static final long serialVersionUID = 1L;
-	private int codigo_loc;
+	private Long codigo_loc;
 	private int cep_loc;
 	private String endere_loc;
 	private String numero_loc;
@@ -22,20 +23,18 @@ public class Localidade implements Serializable {
 	public Localidade() {
 	}
 
-
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name="codigo_loc")
-	public int getCodigo() {
+	@Column(name = "codigo_loc")
+	public Long getCodigo() {
 		return this.codigo_loc;
 	}
 
-	public void setCodigo(int codigo) {
+	public void setCodigo(Long codigo) {
 		this.codigo_loc = codigo;
 	}
 
-
-	@Column(name="cep_loc")
+	@Column(name = "cep_loc")
 	public int getCep() {
 		return this.cep_loc;
 	}
@@ -44,8 +43,7 @@ public class Localidade implements Serializable {
 		this.cep_loc = cep;
 	}
 
-
-	@Column(name="endere_loc")
+	@Column(name = "endere_loc")
 	public String getEndereco() {
 		return this.endere_loc;
 	}
@@ -54,8 +52,7 @@ public class Localidade implements Serializable {
 		this.endere_loc = endereco;
 	}
 
-
-	@Column(name="numero_loc")
+	@Column(name = "numero_loc")
 	public String getNumero() {
 		return this.numero_loc;
 	}
@@ -64,9 +61,8 @@ public class Localidade implements Serializable {
 		this.numero_loc = numero;
 	}
 
-
-	//bi-directional many-to-one association to Estacionamento
-	@OneToMany(mappedBy="localidade")
+	// bi-directional many-to-one association to Estacionamento
+	@OneToMany(mappedBy = "localidade")
 	public List<Estacionamento> getEstacionamento() {
 		return this.estacionamento;
 	}
@@ -89,10 +85,9 @@ public class Localidade implements Serializable {
 		return estacionamento;
 	}
 
-
-	//bi-directional many-to-one association to Bairro
+	// bi-directional many-to-one association to Bairro
 	@ManyToOne
-	@JoinColumn(name="codbai_loc")
+	@JoinColumn(name = "codbai_loc")
 	public Bairro getBairro() {
 		return this.codbai_loc;
 	}
@@ -101,10 +96,9 @@ public class Localidade implements Serializable {
 		this.codbai_loc = bairro;
 	}
 
-
-	//bi-directional many-to-one association to Cidade
+	// bi-directional many-to-one association to Cidade
 	@ManyToOne
-	@JoinColumn(name="codcid_loc")
+	@JoinColumn(name = "codcid_loc")
 	public Cidade getCidade() {
 		return this.codcid_loc;
 	}
